add 8 bit color code and inversion status to TextDecoration.
"default" color, 8 bit color code or 24 bit color is represented by ColorSpec.
Background color and fore color are maintained as ColorSpec.
They are converted to GAttr, and actual color to display is detemined at the time of drawing.
As a result, change of the color table will affect directly to the whole lines.
It is different from previous versions, but it would be more preferred in many cases.

Currently, SGR parameter "2" (faint, decreased intensity) is ignored.